+++ lua/quarto/runner/init.lua
@@ -1,11 +1,10 @@
 --- Code runner, configurable to use different engines.
 local Runner = {}
 
-local otterkeeper = require("otter.keeper")
-local config = require("quarto.config").config
+local otterkeeper = require 'otter.keeper'
+local config = require('quarto.config').config
 
-local no_code_found =
-  "No code chunks found for the current language, which is detected based on the current code block. Is your cursor in a code block?"
+local no_code_found = 'No code chunks found for the current language, which is detected based on the current code block. Is your cursor in a code block?'
 
 local function overlaps_range(range, other)
   return range.from[1] <= other.to[1] and other.from[1] <= range.to[1]
@@ -90,12 +89,12 @@ local function run(range, multi_lang)
   end
 
   otterkeeper.sync_raft(buf)
-  local oa = otterkeeper._otters_attached[buf]
+  local oa = otterkeeper.rafts[buf]
   if oa == nil then
     vim.notify("[Quarto] it seems that the code runner isn't initialized for this buffer.", vim.log.levels.ERROR)
     return
   end
-  local chunks = otterkeeper._otters_attached[buf].code_chunks
+  local chunks = otterkeeper.rafts[buf].code_chunks
 
   local filtered = extract_code_cells_in_range(lang, chunks, range)
 
@@ -166,14 +165,14 @@ Runner.run_range = function()
     return
   end
 
-  local vstart = vim.fn.getpos("'<")
-  local vend = vim.fn.getpos("'>")
+  local vstart = vim.fn.getpos "'<"
+  local vend = vim.fn.getpos "'>"
 
   if vstart and vend then
     local range = { from = { vstart[2] - 1, vstart[1] }, to = { vend[2], vend[1] } }
-    send({ lang = otterkeeper.get_current_language_context(), range = range, text = lines })
+    send { lang = otterkeeper.get_current_language_context(), range = range, text = lines }
   else
-    print("No visual selection")
+    print 'No visual selection'
   end
 end