+++ plugin/quarto.lua
@@ -1,4 +1,4 @@
-if vim.fn.has("nvim-0.9.0") ~= 1 then
+if vim.fn.has 'nvim-0.9.0' ~= 1 then
   local msg = [[
   quarto-dev/quarto-nvim and jmbuhr/otter.nvim require Neovim version >= 0.9.0 (https://github.com/neovim/neovim/releases/tag/stable).
   If you are unable to update Neovim, you can specify a specific version of the plugins involved instead of the latest stable version.
@@ -12,11 +12,10 @@ if vim.fn.has("nvim-0.9.0") ~= 1 then
   return
 end
 
-local quarto = require("quarto")
+local quarto = require 'quarto'
 local api = vim.api
 
-api.nvim_create_user_command("QuartoPreview", quarto.quartoPreview, { nargs = "*" })
-api.nvim_create_user_command("QuartoClosePreview", quarto.quartoClosePreview, {})
-api.nvim_create_user_command("QuartoActivate", quarto.activate, {})
-api.nvim_create_user_command("QuartoHelp", quarto.searchHelp, { nargs = 1 })
-api.nvim_create_user_command("QuartoHover", ':require"otter".ask_hover()<cr>', {})
+api.nvim_create_user_command('QuartoPreview', quarto.quartoPreview, { nargs = '*' })
+api.nvim_create_user_command('QuartoClosePreview', quarto.quartoClosePreview, {})
+api.nvim_create_user_command('QuartoActivate', quarto.activate, {})
+api.nvim_create_user_command('QuartoHelp', quarto.searchHelp, { nargs = 1 })